fix(lobby) don't mix web and native actions on the same file
USe actions.any for common actions.
